Algebra.Algorithms.Groebner.Homogeneous

Documentation

calcGroebnerBasisAfterHomogenising :: (Field (Coefficient poly), IsOrderedPolynomial poly) => Ideal poly -> [poly]

calcGroebnerBasisAfterHomogenisingHilb :: (Field (Coefficient poly), IsOrderedPolynomial poly) => Ideal poly -> [poly]

calcGroebnerBasisAfterHomogenisingWith :: (Field (Coefficient poly), IsOrderedPolynomial poly) => (forall poly'. (Field (Coefficient poly'), IsOrderedPolynomial poly') => Ideal poly' -> [poly']) -> Ideal poly -> [poly]

Calculates Groebner basis once homogenise, apply unsafeCalcHomogeneousGroebnerBasis, and then dehomogenise.

calcHomogeneousGroebnerBasis :: (Field (Coefficient poly), IsOrderedPolynomial poly) => Ideal poly -> Maybe [poly]

Calculates a Groebner basis of the given homogeneous ideal, i.e. an ideal generated by homogeneous polynomials. Returns Nothing if the given ideal is inhomogeneous.

See also unsafeCalcHomogeneousGroebnerBasis.

unsafeCalcHomogeneousGroebnerBasis :: (Field (Coefficient poly), IsOrderedPolynomial poly) => Ideal poly -> [poly]

Calculates a Groebner basis of the given homogeneous ideal, i.e. an ideal generated by homogeneous polynomials.

N.B. This function DOES NOT check homogeniety of the given ideal. See also calcHomogeneousGroebnerBasis.

hilbertPoincareSeries :: (Field (Coefficient poly), IsOrderedPolynomial poly) => Ideal poly -> HPS (Arity poly)

A variant of hilbertPoincareSeriesBy using Grevlex ordering.

hilbertPoincareSeriesBy :: forall ord poly. (IsMonomialOrder (Arity poly) ord, Field (Coefficient poly), IsOrderedPolynomial poly) => ord -> Ideal poly -> HPS (Arity poly)

Calculates the Hilbert-Poincare serires of a given homogeneous ideal, using the specified monomial ordering.

hilbertPoincareSeriesForMonomials :: forall t n. (KnownNat n, Foldable t) => t (Monomial n) -> HPS n

data HPS n

taylorHPS :: HPS n -> [Integer]

toRationalFunction :: KnownNat n => HPS n -> RationalFunction Rational

calcHomogeneousGroebnerBasisHilbert :: (Field (Coefficient poly), IsOrderedPolynomial poly) => Ideal poly -> [poly]

Calculates homogeneous Groebner basis by Hilbert-driven method, computing Hilbert-Poincare series w.r.t. Grevlex.

calcHomogeneousGroebnerBasisHilbertBy :: (Field (Coefficient poly), IsOrderedPolynomial poly, IsMonomialOrder (Arity poly) ord) => ord -> Ideal poly -> [poly]

First compute Hilbert-Poicare series w.r.t. ord by hilbertPoincareSeriesBy, and then apply calcHomogeneousGroebnerBasisHilbertWithSeries.

calcHomogeneousGroebnerBasisHilbertWithSeries :: (Field (Coefficient poly), IsOrderedPolynomial poly) => Ideal poly -> HPS (Arity poly) -> [poly]