Algebra.Algorithms.Groebner.Signature.Rules

Description

This module provides a series of rewriting rules, replacing invocations on calcGroebnerBasis and calcGroebnerBasisWith with f5 followed by minimisation and reduction of basis. Since f5 is the fastest implementation for general Gröbner basis computation in this package, it is expected that doing this improves the performance.

You can import this module to gain a speed-up also in other functions using calcGroebnerBasis such as intersection and so on.

Rules replaces all the occurences of calcGroebnerBasis and calcGroebnerBasisWith with f5. This effect is pervasive once this module is imported; in general, importing this module in library-site is NOT good idea.

In addition, we DO NOT rewrite other Buchberger-specific functions such as calcGroebnerBasisWithStrategy or syzygyBuchberger.